handle not found from grpc error
